Add admin and unauthenticated user header and navbar components; refactor headers and navbars.
