+++ src/app/components/UserHeader.js
@@ -1,19 +1,15 @@
 "use client";
 
 import { Button } from "react-bootstrap"
-import { useRouter } from "next/navigation"
 import { logout } from "../../../utils/supabase/actions";
 
 const Header = () => {
-
-  const router = useRouter()
 
   return (
     <>
     <div>Header</div>
     <h1>Solarize</h1>
-    <Button variant="primary" onClick={() => router.push("/signup")}>Sign Up!</Button>
-    <Button variant="primary" onClick={() => router.push("/login")}>Log In</Button>
+    <h3>Logged in as Resident!</h3>
     <Button variant="danger" onClick={() => logout()}>Log Out</Button>
     </>
   )
